♻️  JAVA-3298 Align Tests to Package Structure
