+++ src/test/java/com/contrastsecurity/http/ApplicationFilterFilterFormTest.java
@@ -1,11 +1,7 @@
-package com.contrastsecurity;
+package com.contrastsecurity.http;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.contrastsecurity.http.ApplicationFilterForm;
-import com.contrastsecurity.http.FilterForm;
-import com.contrastsecurity.http.RuleSeverity;
-import com.contrastsecurity.http.ServerEnvironment;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.EnumSet;
